refactor(spells): Fixed the spell list and page
